+++ changelog.md
@@ -1,8 +1,10 @@
 ==== 1.4.1 ====
 
   * FIXED: Updated Railcraft integration (requires Railcraft 9.5 now)
-  * CHANGED: Some changes regarding railcraft integration to the config file
-    - It is recommended to remove anything with "railcraft" in it from the config and let those entries regenerate
+  * CHANGED: Some changes to the config file
+    - It is recommended to regenerate the config file
+  * CHANGED: Got rid of the EnderIO CapacitorBank's getEnergyStoredForNetwork and getMaxEnergyStoredForNetwork methods
+    - Their behaviour was the same is getEnergyStored/getMaxEnergyStored, so use those methods instead
 
 ==== 1.4.0 ====